Fixes to catalog/bundle generation (#34)
The bundle was being generated with the incorrect operator image
tag which causes a failure in catalog generation.  Since latest
tag is used that tag can not be pushed unless bundle and catalog
generation succeed.  Also changed the bundle/catalog env names to
be more consistent with other names for images