Make validation timer based and disable caching
DVO is caching _a lot_ of different objects throughout a cluster,
leading to huge caches. As DVO does not need to check objects on every
single update and is only accessing each object state once, so the
allocated caches are rather underutilized.

This commit disables caches completely and replaces active watching with
a simple polling job.

Signed-off-by: Nico Schieder <nschieder@redhat.com>